# users-service/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users Service API",
    "description": "Swagger documentation for Users microservice",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Users related endpoints"
    },
    {
      "name": "Goals",
      "description": "Goals related endpoints"
    },
    {
      "name": "Measurements",
      "description": "Measurements related endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "New user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "409": {
            "description": "Email already exists"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/GetUser"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Current user retrieved successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Student details retrieved successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update a User",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/goals/": {
      "post": {
        "tags": [
          "Goals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "New goal object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateGoal"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Goal created successfully",
            "schema": {
              "$ref": "#/definitions/GetGoal"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Goals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Goals found successfully",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/GetGoal"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Goals not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/goals/{id}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Goal found successfully",
            "schema": {
              "$ref": "#/definitions/GetGoal"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Goals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update a Goal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateGoal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goal updated successfully",
            "schema": {
              "$ref": "#/definitions/GetGoal"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Goals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Goal deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/measurements/": {
      "post": {
        "tags": [
          "Measurements"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "New measurement object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateMeasurement"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Measurement created successfully",
            "schema": {
              "$ref": "#/definitions/GetMeasurement"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Measurements found successfully",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/GetMeasurement"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Measurements not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/measurements/{id}": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Measurement found successfully",
            "schema": {
              "$ref": "#/definitions/GetMeasurement"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Measurement not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Measurements"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update a Measurement",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateMeasurement"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Measurement updated successfully",
            "schema": {
              "$ref": "#/definitions/GetMeasurement"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Measurement not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Measurements"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Measurement deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Measurement not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "GetUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Example Name"
        },
        "email": {
          "type": "string",
          "example": "example@mail.com"
        },
        "role": {
          "type": "string",
          "example": "user"
        }
      }
    },
    "CreateUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example Name"
        },
        "email": {
          "type": "string",
          "example": "example@mail.com"
        },
        "password": {
          "type": "string",
          "example": "mypassword123"
        },
        "gender": {
          "type": "string",
          "example": "female"
        },
        "birth_date": {
          "type": "string",
          "example": "1999-05-10"
        },
        "role": {
          "type": "string",
          "example": "user"
        }
      }
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "example@mail.com"
        },
        "password": {
          "type": "string",
          "example": "mypassword123"
        }
      }
    },
    "GetGoal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "user_id": {
          "type": "number",
          "example": 1
        },
        "type": {
          "type": "string",
          "example": "weight_loss"
        },
        "target_value": {
          "type": "number",
          "example": 65
        },
        "unit": {
          "type": "string",
          "example": "kg"
        },
        "target_date": {
          "type": "string",
          "example": "2024-12-31"
        },
        "status": {
          "type": "string",
          "example": "in_progress"
        }
      }
    },
    "CreateGoal": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "weight_loss"
        },
        "target_value": {
          "type": "number",
          "example": 65
        },
        "unit": {
          "type": "string",
          "example": "kg"
        },
        "target_date": {
          "type": "string",
          "example": "2024-12-31"
        },
        "status": {
          "type": "string",
          "example": "in_progress"
        }
      }
    },
    "GetMeasurement": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "user_id": {
          "type": "number",
          "example": 1
        },
        "date": {
          "type": "string",
          "example": "2024-06-01"
        },
        "weight_kg": {
          "type": "number",
          "example": 70
        },
        "height_cm": {
          "type": "number",
          "example": 175
        },
        "body_fat_pct": {
          "type": "number",
          "example": 15.5
        }
      }
    },
    "CreateMeasurement": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "example": "2024-06-01"
        },
        "weight_kg": {
          "type": "number",
          "example": 70
        },
        "height_cm": {
          "type": "number",
          "example": 175
        },
        "body_fat_pct": {
          "type": "number",
          "example": 15.5
        }
      }
    }
  }
}